Data.Dependent.OpenUnion

Description

This implementation relies on _closed_ type families added to GHC 7.8. It has NO overlapping instances and NO Typeable. Alas, the absence of Typeable means the projections and injections generally take linear time. The code illustrate how to use closed type families to disambiguate otherwise overlapping instances.

The data constructors of Union are not exported. Essentially, the nested Either data type.

Using http://okmij.org/ftp/Haskell/extensible/OpenUnion41.hs as a starting point.

Documentation

data Union r v

class Member' t r (FindElem t r) => Member t r where

Associated Types

type Remove t r :: [k -> *]

Methods

weaken :: Proxy t -> Union (Remove t r) v -> Union r v

inj :: t v -> Union r v

prj :: Union r v -> Either (Union (Remove t r) v) (t v)

type family Members m r :: Constraint

Equations

Members (t : c) r = (Member t r, Members c r)

Members `[]` r = ()

hoistU :: forall t r f r' a. (Member t r, Member f r', Remove t r ~ Remove f r') => (t a -> f a) -> Union r a -> Union r' a